requirements.txt
praw==7.8.1
anthropic==0.49.0
notion-client==2.3.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
